se crearon diagramas de paquetes. hubo modificaciones en lasrutas de los archivos
