+++ src/main/java/edu/fiuba/algo3/controladores/ControladorActivarDuplicador.java
@@ -1,21 +1,21 @@
-package edu.fiuba.algo3.controlador;
+package edu.fiuba.algo3.controladores;
 
 import edu.fiuba.algo3.modelo.Panel;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.scene.control.Button;
 
-public class ControladorActivarTriplicador implements EventHandler<ActionEvent> {
+public class ControladorActivarDuplicador implements EventHandler<ActionEvent> {
     private final Panel panel;
     private final Button boton;
 
-    public ControladorActivarTriplicador(Panel panel, Button botonTriplicador) {
+    public ControladorActivarDuplicador(Panel panel, Button botonDuplicador) {
         this.panel = panel;
-        this.boton = botonTriplicador;
+        this.boton = botonDuplicador;
     }
 
     public void handle(ActionEvent actionEvent) {
-        panel.activarTriplicador();
+        panel.activarDuplicador();
         boton.setDisable(true);
     }
 }
