+++ src/main/java/edu/fiuba/algo3/controladores/ControladorIniciarJuego.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.controlador;
+package edu.fiuba.algo3.controladores;
 
 import edu.fiuba.algo3.modelo.Panel;
 import edu.fiuba.algo3.vista.VistaAgregarJugador;
