+++ src/main/java/edu/fiuba/algo3/vista/botones/BotonDuplicador.java
@@ -1,7 +1,7 @@
 package edu.fiuba.algo3.vista.botones;
 
 
-import edu.fiuba.algo3.controlador.ControladorActivarDuplicador;
+import edu.fiuba.algo3.controladores.ControladorActivarDuplicador;
 import edu.fiuba.algo3.modelo.Panel;
 import javafx.scene.control.Button;
