+++ src/main/java/edu/fiuba/algo3/vista/botones/BotonIniciarJuego.java
@@ -1,7 +1,7 @@
 package edu.fiuba.algo3.vista.botones;
 
 
-import edu.fiuba.algo3.controlador.ControladorIniciarJuego;
+import edu.fiuba.algo3.controladores.ControladorIniciarJuego;
 import edu.fiuba.algo3.modelo.Panel;
 import edu.fiuba.algo3.vista.VistaAgregarJugador;
 import javafx.scene.control.Button;
