+++ src/main/java/edu/fiuba/algo3/vista/botones/BotonOpcion.java
@@ -1,6 +1,6 @@
 package edu.fiuba.algo3.vista.botones;
 
-import edu.fiuba.algo3.controlador.ControladorBotonOpcion;
+import edu.fiuba.algo3.controladores.ControladorBotonOpcion;
 import javafx.scene.control.Button;
 
 import java.util.Collection;
