+++ src/main/java/edu/fiuba/algo3/vista/botones/BotonOpcionGC.java
@@ -1,7 +1,7 @@
 package edu.fiuba.algo3.vista.botones;
 
-import edu.fiuba.algo3.controlador.ControladorDragGC;
-import edu.fiuba.algo3.controlador.ControladorDropGC;
+import edu.fiuba.algo3.controladores.ControladorDragGC;
+import edu.fiuba.algo3.controladores.ControladorDropGC;
 import javafx.scene.control.Button;
 
 public class BotonOpcionGC extends Button {
