+++ src/main/java/edu/fiuba/algo3/vista/botones/BotonResponder.java
@@ -1,7 +1,7 @@
 package edu.fiuba.algo3.vista.botones;
 
-import edu.fiuba.algo3.Timer;
-import edu.fiuba.algo3.controlador.ControladorResponder;
+import edu.fiuba.algo3.vista.Timer;
+import edu.fiuba.algo3.controladores.ControladorResponder;
 import edu.fiuba.algo3.modelo.Panel;
 import javafx.scene.control.Button;
