+++ src/main/java/edu/fiuba/algo3/vista/botones/BotonResponderGC.java
@@ -1,8 +1,8 @@
 package edu.fiuba.algo3.vista.botones;
 
 
-import edu.fiuba.algo3.Timer;
-import edu.fiuba.algo3.controlador.ControladorResponderGC;
+import edu.fiuba.algo3.vista.Timer;
+import edu.fiuba.algo3.controladores.ControladorResponderGC;
 import edu.fiuba.algo3.modelo.Panel;
 import javafx.scene.control.Button;
