+++ src/main/java/edu/fiuba/algo3/vista/botones/BotonTriplicador.java
@@ -1,6 +1,6 @@
 package edu.fiuba.algo3.vista.botones;
 
-import edu.fiuba.algo3.controlador.ControladorActivarTriplicador;
+import edu.fiuba.algo3.controladores.ControladorActivarTriplicador;
 import edu.fiuba.algo3.modelo.Panel;
 import javafx.scene.control.Button;
