+++ src/main/java/edu/fiuba/algo3/vista/cajas/CajaTimer.java
@@ -1,6 +1,6 @@
 package edu.fiuba.algo3.vista.cajas;
 
-import edu.fiuba.algo3.Timer;
+import edu.fiuba.algo3.vista.Timer;
 import edu.fiuba.algo3.modelo.Panel;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
