+++ src/main/java/edu/fiuba/algo3/vista/vistaPregunta/VistaPregunta.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.vista;
+package edu.fiuba.algo3.vista.vistaPregunta;
 
 
 import javafx.scene.Scene;
