+++ src/main/java/edu/fiuba/algo3/vista/vistaPregunta/VistaPreguntaClasica.java
@@ -1,7 +1,9 @@
-package edu.fiuba.algo3.vista;
+package edu.fiuba.algo3.vista.vistaPregunta;
 
 import edu.fiuba.algo3.modelo.Panel;
-import edu.fiuba.algo3.vista.cajas.*;
+import edu.fiuba.algo3.vista.cajas.CajaJugadores;
+import edu.fiuba.algo3.vista.cajas.CajaPregunta;
+import edu.fiuba.algo3.vista.cajas.CajaTimer;
 import javafx.geometry.Insets;
 import javafx.scene.Scene;
 import javafx.scene.layout.BorderPane;
@@ -9,29 +11,31 @@ import javafx.scene.layout.BorderPane;
 import java.util.HashSet;
 import java.util.Set;
 
-public class VistaPreguntaConPenalidad implements VistaPregunta {
+public class VistaPreguntaClasica implements VistaPregunta {
+
     private final Panel panel;
     private final CajaJugadores cajaJugadores;
     private final CajaTimer cajaTimer;
 
-    public VistaPreguntaConPenalidad(Panel panel, CajaJugadores cajaJugadores, CajaTimer cajaTimer) {
+    public VistaPreguntaClasica(Panel panel, CajaJugadores cajaJugadores, CajaTimer cajaTimer) {
         this.panel = panel;
         this.cajaJugadores = cajaJugadores;
         this.cajaTimer = cajaTimer;
     }
 
-    @Override
+
     public Scene devolverVistaPregunta() {
+
         BorderPane componentLayout = new BorderPane();
         componentLayout.setPadding(new Insets(20, 100, 10, 100));
 
         Set<String> respuestasJugador = new HashSet<>();
 
-
         CajaPregunta cajaPregunta = new CajaPregunta(panel, respuestasJugador, cajaTimer.devolverReloj());
+
         panel.obtenerPreguntaActual().agregarObservador(cajaPregunta);
 
-        cajaJugadores.activarMultiplicadores();
+        cajaJugadores.activarExclusividades();
 
         componentLayout.setCenter(cajaPregunta);
         componentLayout.setBottom(cajaJugadores);
