+++ src/main/java/edu/fiuba/algo3/vista/vistaPregunta/VistaPreguntaConPenalidad.java
@@ -1,9 +1,7 @@
-package edu.fiuba.algo3.vista;
+package edu.fiuba.algo3.vista.vistaPregunta;
 
 import edu.fiuba.algo3.modelo.Panel;
-import edu.fiuba.algo3.vista.cajas.CajaJugadores;
-import edu.fiuba.algo3.vista.cajas.CajaPregunta;
-import edu.fiuba.algo3.vista.cajas.CajaTimer;
+import edu.fiuba.algo3.vista.cajas.*;
 import javafx.geometry.Insets;
 import javafx.scene.Scene;
 import javafx.scene.layout.BorderPane;
@@ -11,31 +9,29 @@ import javafx.scene.layout.BorderPane;
 import java.util.HashSet;
 import java.util.Set;
 
-public class VistaPreguntaClasica implements VistaPregunta {
-
+public class VistaPreguntaConPenalidad implements VistaPregunta {
     private final Panel panel;
     private final CajaJugadores cajaJugadores;
     private final CajaTimer cajaTimer;
 
-    public VistaPreguntaClasica(Panel panel, CajaJugadores cajaJugadores, CajaTimer cajaTimer) {
+    public VistaPreguntaConPenalidad(Panel panel, CajaJugadores cajaJugadores, CajaTimer cajaTimer) {
         this.panel = panel;
         this.cajaJugadores = cajaJugadores;
         this.cajaTimer = cajaTimer;
     }
 
-
+    @Override
     public Scene devolverVistaPregunta() {
-
         BorderPane componentLayout = new BorderPane();
         componentLayout.setPadding(new Insets(20, 100, 10, 100));
 
         Set<String> respuestasJugador = new HashSet<>();
 
-        CajaPregunta cajaPregunta = new CajaPregunta(panel, respuestasJugador, cajaTimer.devolverReloj());
 
+        CajaPregunta cajaPregunta = new CajaPregunta(panel, respuestasJugador, cajaTimer.devolverReloj());
         panel.obtenerPreguntaActual().agregarObservador(cajaPregunta);
 
-        cajaJugadores.activarExclusividades();
+        cajaJugadores.activarMultiplicadores();
 
         componentLayout.setCenter(cajaPregunta);
         componentLayout.setBottom(cajaJugadores);
