+++ src/main/java/edu/fiuba/algo3/vista/vistaPregunta/VistaPreguntaOC.java
@@ -1,38 +1,38 @@
-package edu.fiuba.algo3.vista;
+package edu.fiuba.algo3.vista.vistaPregunta;
 
 import edu.fiuba.algo3.modelo.Panel;
-import edu.fiuba.algo3.vista.cajas.*;
+import edu.fiuba.algo3.vista.cajas.CajaJugadores;
+import edu.fiuba.algo3.vista.cajas.CajaPregunta;
+import edu.fiuba.algo3.vista.cajas.CajaTimer;
 import javafx.geometry.Insets;
 import javafx.scene.Scene;
 import javafx.scene.layout.BorderPane;
 
-import java.util.HashSet;
-import java.util.Set;
+import java.util.LinkedList;
+import java.util.List;
 
-public class VistaPreguntaGC implements VistaPregunta {
+public class VistaPreguntaOC implements VistaPregunta {
     private final Panel panel;
     private final CajaJugadores cajaJugadores;
     private final CajaTimer cajaTimer;
 
-    public VistaPreguntaGC(Panel panel, CajaJugadores cajaJugadores, CajaTimer cajaTimer) {
+    public VistaPreguntaOC(Panel panel, CajaJugadores cajaJugadores, CajaTimer cajaTimer) {
         this.panel = panel;
         this.cajaJugadores = cajaJugadores;
         this.cajaTimer = cajaTimer;
     }
 
-
+    @Override
     public Scene devolverVistaPregunta() {
-
         BorderPane componentLayout = new BorderPane();
         componentLayout.setPadding(new Insets(20, 100, 10, 100));
 
-        Set<String> respuestasJugador = new HashSet<>();
-        respuestasJugador.add(panel.obtenerPreguntaActual().obtenerGrupoCorrecto());
+        List<String> respuestasJugador = new LinkedList<>();
+
+        CajaPregunta cajaPregunta = new CajaPregunta(panel, respuestasJugador, cajaTimer.devolverReloj());
+        panel.obtenerPreguntaActual().agregarObservador(cajaPregunta);
 
         cajaJugadores.activarExclusividades();
-
-        CajaPreguntaGC cajaPregunta = new CajaPreguntaGC(panel, respuestasJugador, cajaTimer.devolverReloj());
-        panel.obtenerPreguntaActual().agregarObservador(cajaPregunta);
 
         componentLayout.setCenter(cajaPregunta);
         componentLayout.setBottom(cajaJugadores);
@@ -40,4 +40,5 @@ public class VistaPreguntaGC implements VistaPregunta {
 
         return new Scene(componentLayout, 640, 480);
     }
+
 }
